+++ config/routes.rb
@@ -113,6 +113,12 @@ Rails.application.routes.draw do
           get "completed", to: "projects#completed"
           get "unassigned", to: "projects#unassigned"
         end
+        namespace :regional, path: "regional" do
+          get "in-progress", to: "projects#in_progress"
+          get "completed", to: "projects#completed"
+          get ":region/in-progress", to: "projects#in_progress_by_region", as: :in_progress_by_region
+          get ":region/completed", to: "projects#completed_by_region", as: :completed_by_region
+        end
         namespace :user do
           get "in-progress", to: "projects#in_progress"
           get "completed", to: "projects#completed"